⛳fix: price delta in OlympusPrice.updateMovingAverage() can be unchecked
C44 - Use unchecked blocks where overflow/underflow is impossible, or where an overflow/underflow is unrealistic on human timescales (counters, etc). Comment explanations wherever unchecked is used, along with an estimate of how much gas it saves (if relevant).

NOTE: the rest of the math in the conditional can/should be unchecked to not only save gas but to prevent DOS in the basically impossible case that the moving average is too large or small.

https://github.com/transmissions11/solcurity